+++ src/representative_subset_calculator/similarity_matrix/similarity_matrix.h
@@ -7,13 +7,13 @@
 
 class SimilarityMatrix {
     public:
-    virtual double getCoverage() const = 0; 
+    virtual float getCoverage() const = 0; 
 };
 
 class MutableSimilarityMatrix : public SimilarityMatrix {
     private:
     std::vector<const DataRow*> baseRows;
-    std::vector<std::vector<double>> transposeMatrix;
+    std::vector<std::vector<float>> transposeMatrix;
 
     public:
     MutableSimilarityMatrix() {}
@@ -32,12 +32,12 @@ class MutableSimilarityMatrix : public SimilarityMatrix {
         return this->baseRows;
     }
 
-    const std::vector<std::vector<double>> &getTranspose() const {
+    const std::vector<std::vector<float>> &getTranspose() const {
         return this->transposeMatrix;
     }
 
     void addRow(const DataRow &newRow) {
-        this->transposeMatrix.push_back(std::vector<double>(this->transposeMatrix.size() + 1));
+        this->transposeMatrix.push_back(std::vector<float>(this->transposeMatrix.size() + 1));
         for (size_t i = 0; i < this->baseRows.size(); i++) {
             this->transposeMatrix[i].push_back(this->baseRows[i]->dotProduct(newRow));
             this->transposeMatrix.back()[i] = this->transposeMatrix[i].back();
@@ -47,15 +47,15 @@ class MutableSimilarityMatrix : public SimilarityMatrix {
         this->transposeMatrix.back().back() = newRow.dotProduct(newRow) + 1;
     }
 
-    double getCoverage() const {
-        Eigen::MatrixXd kernelMatrix(this->transposeMatrix.size(), this->transposeMatrix.size());
+    float getCoverage() const {
+        Eigen::MatrixXf kernelMatrix(this->transposeMatrix.size(), this->transposeMatrix.size());
         for (int i = 0; i < this->transposeMatrix.size(); i++) {
-            kernelMatrix.row(i) = Eigen::VectorXd::Map(this->transposeMatrix[i].data(), this->transposeMatrix[i].size());
+            kernelMatrix.row(i) = Eigen::VectorXf::Map(this->transposeMatrix[i].data(), this->transposeMatrix[i].size());
         }
 
-        Eigen::MatrixXd diagonal(kernelMatrix.llt().matrixL());
+        Eigen::MatrixXf diagonal(kernelMatrix.llt().matrixL());
 
-        double res = 0;
+        float res = 0;
         for (size_t index = 0; index < diagonal.rows(); index++) {
             res += std::log(diagonal(index,index));
         }
